+++ .circleci/config.yml
@@ -15,9 +15,9 @@ jobs:
       # based on this: https://discuss.circleci.com/t/circle-2-0-caching-is-too-limited-to-be-very-useful/11694/10
       - restore_cache:
           keys:
-            - spring-boot-crud-booster-{{ .Branch }}-{{ checksum "pom.xml" }}
-            - spring-boot-crud-booster-{{ .Branch }}
-            - spring-boot-crud-booster
+            - crud-example-{{ .Branch }}-{{ checksum "pom.xml" }}
+            - crud-example-{{ .Branch }}
+            - crud-example
 
       - run:
           name: Install Maven dependencies locally
@@ -45,7 +45,7 @@ jobs:
       - save_cache:
           paths:
             - ~/.m2
-          key: spring-boot-crud-booster-{{ .Branch }}-{{ checksum "pom.xml" }}
+          key: crud-example-{{ .Branch }}-{{ checksum "pom.xml" }}
 
       - run:
           name: run tests against s2i build
@@ -66,7 +66,7 @@ jobs:
             oc create -f .openshiftio/service.yaml
 
             oc create -f .openshiftio/application.yaml
-            oc new-app --template=spring-boot-rest-http-crud -p SOURCE_REPOSITORY_URL=${SOURCE_REPOSITORY_URL} -p SOURCE_REPOSITORY_REF=${SOURCE_REPOSITORY_REF}
+            oc new-app --template=crud -p SOURCE_REPOSITORY_URL=${SOURCE_REPOSITORY_URL} -p SOURCE_REPOSITORY_REF=${SOURCE_REPOSITORY_REF}
 
             sleep 30 # needed in order to bypass the 'Pending' state
             # wait for the app to stand up
