+++ .openshiftio/application.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Template
 metadata:
-  name: spring-boot-rest-http-crud
+  name: crud
   annotations:
     iconClass: icon-spring
     tags: spring-boot, rest, crud, java, microservice
@@ -9,7 +9,7 @@ metadata:
     openshift.io/provider-display-name: "Red Hat, Inc."
     openshift.io/documentation-url: "https://appdev.prod-preview.openshift.io/docs/spring-boot-runtime.html#mission-crud-spring-boot-tomcat"
     description: >-
-      The Relational Database Backend booster expands on the REST API Level 0 booster to provide a basic example of performing create, read, update and delete (CRUD) operations on a PostgreSQL database using a simple HTTP API.
+      The Relational Database Backend example expands on the REST API Level 0 example to provide a basic example of performing create, read, update and delete (CRUD) operations on a PostgreSQL database using a simple HTTP API.
       CRUD operations are the four basic functions of persistent storage, widely used when developing an HTTP API dealing with a database.
 parameters:
 - name: RUNTIME_VERSION
@@ -50,7 +50,7 @@ objects:
 - apiVersion: v1
   kind: ImageStream
   metadata:
-    name: spring-boot-rest-http-crud
+    name: crud
   spec: {}
 - apiVersion: v1
   kind: ImageStream
@@ -65,12 +65,12 @@ objects:
 - apiVersion: v1
   kind: BuildConfig
   metadata:
-    name: spring-boot-rest-http-crud
+    name: crud
   spec:
     output:
       to:
         kind: ImageStreamTag
-        name: spring-boot-rest-http-crud:BOOSTER_VERSION
+        name: crud:BOOSTER_VERSION
     postCommit: {}
     resources: {}
     source:
@@ -107,10 +107,10 @@ objects:
   kind: Secret
   metadata:
     labels:
-      app: spring-boot-rest-http-crud
+      app: crud
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: my-database-secret
   stringData:
     user: luke
@@ -119,11 +119,11 @@ objects:
   kind: Service
   metadata:
     labels:
-      app: spring-boot-rest-http-crud
+      app: crud
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-rest-http-crud
+      group: dev.snowdrop.example
+    name: crud
   spec:
     ports:
     - name: http
@@ -131,25 +131,25 @@ objects:
       protocol: TCP
       targetPort: 8080
     selector:
-      app: spring-boot-rest-http-crud
+      app: crud
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
     labels:
-      app: spring-boot-rest-http-crud
+      app: crud
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-rest-http-crud
+      group: dev.snowdrop.example
+    name: rest-crud-example
   spec:
     replicas: 1
     revisionHistoryLimit: 2
     selector:
-      app: spring-boot-rest-http-crud
+      app: crud
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -157,10 +157,10 @@ objects:
     template:
       metadata:
         labels:
-          app: spring-boot-rest-http-crud
+          app: crud
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -180,7 +180,7 @@ objects:
             valueFrom:
               fieldRef:
                 fieldPath: metadata.namespace
-          image: spring-boot-rest-http-crud:BOOSTER_VERSION
+          image: crud:BOOSTER_VERSION
           imagePullPolicy: IfNotPresent
           livenessProbe:
             httpGet:
@@ -212,20 +212,20 @@ objects:
         - spring-boot
         from:
           kind: ImageStreamTag
-          name: spring-boot-rest-http-crud:BOOSTER_VERSION
+          name: crud:BOOSTER_VERSION
       type: ImageChange
 - apiVersion: v1
   kind: Route
   metadata:
     labels:
-      app: spring-boot-rest-http-crud
+      app: crud
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-rest-http-crud
+      group: dev.snowdrop.example
+    name: crud
   spec:
     port:
       targetPort: 8080
     to:
       kind: Service
-      name: spring-boot-rest-http-crud
+      name: crud
