+++ README.md
@@ -1,3 +1,3 @@
-[![CircleCI](https://circleci.com/gh/snowdrop/spring-boot-crud-booster.svg?style=shield)](https://circleci.com/gh/snowdrop/spring-boot-crud-booster)
+[![CircleCI](https://circleci.com/gh/snowdrop/rest-crud-example.svg?style=shield)](https://circleci.com/gh/snowdrop/rest-crud-example)
 
 https://appdev.openshift.io/docs/spring-boot-runtime.html#mission-crud-spring-boot
