+++ pom.xml
@@ -21,11 +21,11 @@
     <artifactId>booster-parent</artifactId>
     <version>23</version>
   </parent>
-  <groupId>io.openshift.booster</groupId>
-  <artifactId>spring-boot-rest-http-crud</artifactId>
+  <groupId>dev.snowdrop.example</groupId>
+  <artifactId>crud</artifactId>
   <version>1.5.19-3-SNAPSHOT</version>
-  <name>Spring Boot - CRUD Booster</name>
-  <description>Spring Boot - CRUD Booster</description>
+  <name>Spring Boot - CRUD Example</name>
+  <description>Spring Boot - CRUD Example</description>
   <properties>
     <arquillian.version>1.4.0.Final</arquillian.version>
     <awaitility.version>3.1.0</awaitility.version>
