+++ src/main/java/dev/snowdrop/example/ExampleApplication.java
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package io.openshift.booster;
+package dev.snowdrop.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BoosterApplication {
+public class ExampleApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(BoosterApplication.class, args);
+        SpringApplication.run(ExampleApplication.class, args);
     }
 
 }
