+++ src/main/java/dev/snowdrop/example/exception/UnprocessableEntityException.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.exception;
+package dev.snowdrop.example.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.UNSUPPORTED_MEDIA_TYPE)
-public class UnsupportedMediaTypeException extends RuntimeException {
+@ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
+public class UnprocessableEntityException extends RuntimeException {
 
-    public UnsupportedMediaTypeException(String message) {
+    public UnprocessableEntityException(String message) {
         super(message);
     }
