+++ src/main/java/dev/snowdrop/example/exception/UnsupportedMediaTypeException.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.exception;
+package dev.snowdrop.example.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
-public class UnprocessableEntityException extends RuntimeException {
+@ResponseStatus(HttpStatus.UNSUPPORTED_MEDIA_TYPE)
+public class UnsupportedMediaTypeException extends RuntimeException {
 
-    public UnprocessableEntityException(String message) {
+    public UnsupportedMediaTypeException(String message) {
         super(message);
     }
