+++ src/main/java/dev/snowdrop/example/service/Fruit.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
